WIP: debugging runtime errors: testing editView, want to see if the view inflated for new ingredients behaves how I think it will.

+++ RecipeBinderApp/src/main/java/com/jeffbrandon/recipebinder/room/RecipeDao.kt
@@ -16,7 +16,7 @@ interface RecipeDao {
     fun insertRecipes(recipes: List<RecipeData>): List<Long>
 
     @Query("SELECT * FROM RecipeData WHERE id = :id")
-    fun fetchRecipe(id: Int): RecipeData
+    fun fetchRecipe(id: Long): RecipeData
 
     @Query("SELECT name FROM RecipeData")
     fun fetchRecipeNames(): List<String>
